newer compilers discourage arithmetic comparisons directly between a floating‑point type (double) and an enum

+++ src/programs/AmplitudeAnalysis/project_moments_SPD_etapi0_posepsilon/project_moments_SPD_etapi0_posepsilon.cc
@@ -54,8 +54,8 @@ int main( int argc, char* argv[] ){
         exit(1);
     }
     
-    double step = ( highMass - lowMass ) / kNumBins;
-    double stept = ( hight - lowt ) / kNumBinst;
+    double step = ( highMass - lowMass ) / static_cast<double>(kNumBins);
+    double stept = ( hight - lowt ) / static_cast<double>(kNumBinst);
     
     ofstream outfile;
     outfile.open( outfileName.c_str() );
